backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
marshmallow==3.20.1
google-generativeai==0.3.2
pytest==7.4.3
python-dotenv==1.0.0
psycopg2-binary==2.9.11
